% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kf_add_metric.R
\name{kf_add_metric}
\alias{kf_add_metric}
\title{Add metric to pipeline metrics set}
\usage{
kf_add_metric(metric_set, name, value, format)
}
\arguments{
\item{metric_set}{Metric set from \code{\link[=kf_init_metrics]{kf_init_metrics()}} or a previious call of this function.}

\item{name}{Name of metric.}

\item{value}{Value of metric. Must be numeric.}

\item{format}{Format to display metric in Kubeflow. Must be either RAW or PERCENTAGE.}
}
\value{
Metric set
}
\description{
Adds on to an existing metric set and checks values are formatted correctly.
}
\details{
You can repeatedly chain this function to iteratively build up a pipeline metrics output.
}
\examples{
# Create ROC AUC and PR AUC output
kf_init_metrics() \%>\%
  kf_add_metric("roc-auc", .75, "RAW") \%>\%
  kf_add_metric("pr-auc", .45, "RAW")
}
