% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kf_add_roc.R
\name{kf_add_roc}
\alias{kf_add_roc}
\title{Add a ROC Curve to a UI Metadata Set}
\usage{
kf_add_roc(ui_set, roc_file)
}
\arguments{
\item{ui_set}{UI set from \code{\link[=kf_init_ui_meta]{kf_init_ui_meta()}} or a previious call of this type of function.}

\item{roc_file}{Location of coresponding .csv that contains thresholds, FPR, and TPR.}
}
\value{
UI set
}
\description{
Add a ROC Curve to a UI Metadata Set
}
\examples{
kf_init_ui_meta() \%>\%
  kf_add_roc("gs://project-bucket/roc_file.csv")
}
