% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kf_init_metrics.R
\name{kf_init_metrics}
\alias{kf_init_metrics}
\title{Initialize Pipeline Metrics Artifacts}
\usage{
kf_init_metrics()
}
\value{
A list
}
\description{
Lay the foundational elements of the JSON expected by Kubeflow to provide metrics from your components.
}
\examples{
metrics <- kf_init_metrics()

}
