% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kf_init_ui_meta.R
\name{kf_init_ui_meta}
\alias{kf_init_ui_meta}
\title{Initialize UI Meta Artifacts}
\usage{
kf_init_ui_meta()
}
\value{
A list
}
\description{
Lay the foundational elements of the JSON expected by Kubeflow to provide UI meta data from your components.
}
\examples{
metrics <- kf_init_ui_meta()
kf_add_metric(metrics, "roc", .5, "RAW")
}
