% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kf_write_output.R
\name{kf_write_output}
\alias{kf_write_output}
\title{Write output to a provided outputPath}
\usage{
kf_write_output(contents, outfile)
}
\arguments{
\item{contents}{R object to save.}

\item{outfile}{File provided by Kubeflow}
}
\value{
outfile
}
\description{
Write output to a provided outputPath
}
\examples{
predictions_table_name <- "model_predictions"
mock_path <- tempfile()
kf_write_output(predictions_table_name, mock_path)
}
