% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kflow-package.R
\docType{package}
\name{kflow-package}
\alias{kflow}
\alias{kflow-package}
\title{kflow: Functions for Interacting with Kubeflow from R}
\description{
Convenience functions for writing R code that works in a Kubeflow environment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ndiquattro/kflow}
  \item Report bugs at \url{https://github.com/ndiquattro/kflow/issues}
}

}
\author{
\strong{Maintainer}: Nick DiQuattro \email{ndiquattro@gmail.com}

}
\keyword{internal}
