% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kf_init_metrics.R
\name{print.kflow_meta}
\alias{print.kflow_meta}
\title{Prints kflow_meta objects}
\usage{
\method{print}{kflow_meta}(x, preview = TRUE, pretty = TRUE, ...)
}
\arguments{
\item{x}{kflow_meta object. Likely from \code{\link[=kf_init_metrics]{kf_init_metrics()}}.}

\item{preview}{Boolean to indicate whether JSON should be printed. Set to \code{FALSE} to inspect underlying list structure.}

\item{pretty}{Boolean to indicate whether to pretty print or not.}

\item{...}{Other arguments passed on.}
}
\description{
Extends print to work with kflow_meta objects
}
\examples{
# By default it displays the eventual JSON
test_mets <- kf_init_metrics()
test_mets

# Use print() to see in list form
print(test_mets, preview = FALSE)
}
