% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kf_write_output.R
\name{kf_write_output}
\alias{kf_write_output}
\title{Write output to a provided outputPath}
\usage{
kf_write_output(contents, outfile)
}
\arguments{
\item{contents}{R object to save.}

\item{outfile}{File provided by Kubeflow}
}
\value{
outfile
}
\description{
Writes content to a Kubeflow provided path for output.
}
\details{
Defaults to \code{\link[readr:write_file]{readr::write_file()}}, but uses \code{\link[jsonlite:write_json]{jsonlite::write_json()}} or \code{\link[readr:write_csv]{readr::write_csv()}} for appropriate classes.
}
\examples{
predictions_table_name <- "model_predictions"
mock_path <- tempfile()
kf_write_output(predictions_table_name, mock_path)
}
